backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.7
passlib==1.7.4
pyjwt==2.6.0
motor==3.5.2
pymongo
requests==2.28.2
openai==0.27.0
python-dotenv==0.21.1
email-validator==1.3.1
bcrypt==4.0.1